move file "Terminal"
